new release 25 March 2010 (R4), from 10 August 2010, aka #124
Below is a summary of the new things.

1. Michio Katouda of IMS in Okazaki sent us his RI-MP2 energy
program.  This is able to select between storage of data
in memory or on disk, and so is quite flexible to the computer
hardware you are running on.  The RI method greatly reduces the
amount of data stored, and thus the amount of computation required,
so the timings are very fast.  This is 1032 AOs, with 62 frozen
cores, and 164 valence orbitals correlated (taxol, 6-31G*):
         guess  RHF  props   MP2   total CPU,wall
   p= 1:  33   7391    19   7919    15367, 15367
   p= 2:  21   3694    10   4131     7860,  7975
   p= 4:  19   1857     6   2290     4174,  4184
   p= 8:  30    952     7   1487     2479,  2483    = just 41 minutes!
   p=16:  25    486     5    758     1276,  1341 (two 8-ways/Infiniband)
The results are
   RHF= -2911.8220150649, 16 iters,    E(MP2)= -2920.6062307827
   RHF= -2911.8220150648, 16 iters, E(RI-MP2)= -2920.6075123496
so the correlation energy is good to about 1.5 millihartree error
out of 8.8 Hartree total.  There are inputs to allow bigger auxiliary
bases to reduce the errors.  This program is for the energy only,
but since it can do the energy in about 2x the SCF time, it is certainly
competitive with DFT.  More information is at
  M.Katouda, S.Nagase Int.J.Quantum Chem. 109, 2121-2130(2009)
The input for this program is $MP2 CODE=RIMP2, and then two new
groups, namely $RIMP2 which defaults to storing everything in memory,
both shared and distributed memory.  If you want to use disks instead,
see the inputs in $RIMP2.  The $AUXBAS input is required, to name the
RI basis set.  Thus
 $mp2    code=rimp2 $end
 $auxbas cabnam=svp $end
may be enough input.  The paper mentioned has a calculation out to
4,000 AOs, so it can go really big.

See $MP2 for a comparison of the numerous MP2 programs we now have,
I have brought this up to date.

2. Noriyuki has coded TD-DFT/EFP1 gradients (closed shell reference).
There isn't any new input for this, except that you can now
select RUNTYP=OPTIMIZE (etc) that need gradients.

3. Tamm/Dancoff approximation can be applied to RHF-reference TD-DFT
energies.  This is by Federico, and it doesn't apply to gradients,
nor to UHF.  A proof that Tamm/Dancoff reduces to CIS in the TD-HF
limit can be seen at
       /u1/mike/tests/tddft/acetaldehyde.inp
The input for this is TAMMD=.TRUE. in $TDDFT.

4. There is an FMO update, to make it version 4.0.  Dmitri has
told me the following about this:

FMO gradient improvement (HOP gradient) by T. Nagata. It is strictly
derived for RHF/DFT. For other cases it may be incomplete. i.e., more
terms may be needed, although those present should be pertinent too.
T. Nagata et al., Chem. Phys. Lett. 492 (2010) 302-308.

     (This does not seem to be the same thing as what Kurt
      talked about at group meeting this week)

Screening of ESP based on point charges.
D. G. Fedorov et al, J. Phys. Chem A, in press. DOI: 10.1021/jp101724p

FMO/PCM was made to work with FIXPVA (H. Li).

Some improvements for finding the bond orbital in the model calculations of
FMO/AFO, and a new option to avoid convergence problems with diffuse functions
for these model calculations.

FMO now prints FMO results. This is ridiculous, as it implies. The two F have
a different meaning, the latter one is Frontier (maybe, Fukui, if you are
on the right side of the sea). Note that AE is derived from a "theorem"
(i.e., a poor guess) Koopmans never proposed but it is sometimes imagined as
his anyway).

and that's the end of his summary.

There's no information from Tsukuba about any input changes involved,
except that I know that setting MODGRD=18 in $fmo causes the gradient
to omit the new terms, and thus give old gradients.  I've been promised
documentation at the end of August, when I will not be here.

5. Tinker has been changed to allow more atoms and run in parallel,
this is also to do with FMO calculations.  The storage for the hessian
matrix is now allocated dynamically from MWORDS, rather than being a
compile time limit.  However, the program requires that you give it an
estimate of the hessian storage needed, as MAXHES in $TOPTMZ.  I found
that the size MAXHES=400,000 covered a Si15H16 embedded in Si199H92
cluster.  There are several matrices of this dimension, so MWORDS has
to go up more than this size guess.

There are several new parameter files, saved into a new subdirectory,
such as /clusters/mike/gamess/tinker/params51.  The old parameter files
are completly unchanged, still as subdirectory "params".  The "params51"
directory has a file named 0README to say what the new ones are.

The results for the embedded silicon cluster test case I have are
completely unchanged by this, and the new Tinker is compiled into
the binary on ibmcluster, as usual.

6. Melissa Gajewski who is a student at U. Alberta added the RM1
semi-empirical model.  This is a set of parameters designed to improve
the accuracy over AM1 and PM3, without introducing anything except
new numerical values for existing parameters.  This is selected by
 $basis gbasis=rm1 $end
If you add the variable NGAUSS=X to a semi-empirical run, namely
gbasis=am1, pm3, or rm1, then the log file will contain STO-XG
expansions of the Slater-type-orbital basis in terms of Gaussian.
That permits plotting semi-empirical orbitals with MacMolPlt.

More about RM1 parameters: RM1 is the most recent parameterization,
replacing AM1 data for H, C-F, P-Cl, Br, and I.  See
   G. Bruno Rocha, R. Oliveira Freire, A. Mayall Simas, and
   J. J. P. Stewart, J. Comput. Chem. 27, 1101-1111(2006).

7. Toby Zeng sent a change to permit density matrix at the spin-orbit
coupled level, see ISTNO in $TRANST, and
      /u1/mike/tests/relwfn/bi-j-jz.inp

Fixes since (R3):

One or two of the MP2 programs wasn't quite working with COSMO in
the last code.  Roberto Peverati has fixed all of the PT gradient
codes to work, so COSMO goes with RHF/UHF/ROHF, any DFT counterpart,
and any MP2 counterpart (except the RI-MP2 and OSPT=RMP programs,
which only do PT energy values).

Jerry and his infamous S4 group molecule (same point group as a baseball,
if you include the seams) turned up an error in parallel MCP gradients.
That is fixed by properly zeroing on all nodes.

so, not so many bugs fixed.

The files changed Wednesday numbered 104, plus five new ones:
      aldeci algnci ccsdt ceeis chgpen cisgrd cpmchf
      dccc dcmp2 dcscf dctran ddilib demrpt dftdis
      drc efdrvr efgrdc efinp efinta efintb efpaul
      efpcm efpcov efpfmo elgscf eomcc ffield fmo
      fmocp fmoesd fmogrd fmohss fmoint fmoio fmolib
      fmopbc fmoprp fsodci g3 gamess globop gmcpt
      grd1 grd2a guess gugdga gugdm gugdm2 gugdrt
      gvb hess inputa inputb int1 int2a iolib
      local mccas mcjac mcpgrd mcpinp mcqud mcscf
      mdefp morokm mp2 mp2ddi mp2gr2 mp2grd mp2ims
      mpcdat mpcmol mthlib ormas1 parley pcm pcmcav
      pcmcv2 pcmief prpel prplib prppop qmmm raman
      rhfuhf ricab riint rimp2 rohfcc rxncrd scflib
      scrf sozeff statpt symorb symslc tddefp tddft
      tddfxc tddgrd tdhf trans trnstn umpddi utddft
      vibanl vvos zapddi zmatrx
The new ones in that list are fmocp, fmohss, ricab, riint, rimp2.