Checktest implemented through Makefile.in
The scripts runtest and checktst in tests/ still work and can be used if
the tests folder is copied elsewhere.

The comments in Makefile.in describe how to add new tests to the
checktest target.   The supplied rungms script has been updated to run
without a priori knownledge of the system.  It will inherit information
from install.info to determine the paths and executable name.