LocalGMS used by Makefile.in to run tests
The sockets-only script is now called tools/localgms
and a copy of the released version of rungms has been
placed in the top directory.

Both rungms and localgms get file assignments from
$GMS_PATH/gms-files.csh
and the assignments can be overridden with $HOME/gmsrc
using the CSH setenv command.  For example, to save the
CC amplitudes:

setenv  CCREST $WORKDIR/$JOB.ccrest

The Makefile is now able to run exam inputs with an arbitrary
number of subdirectories since it exports GMS_PATH and
GMS_BUILD_DIR to the shell.  Exams can be run in parallel
within the localhost only:

make test EXAMS=exam01 NCPUS=2

Change-Id: I9b4dbe38f5906689bd05fe3a0d89ab1d41a13848