The Cluster-in-Molecules method
A serial implementation of CIM by Piotr Piecuch and Wei Li
is included here.  A procedure to run the correlated subsystems
in parallel is explained in REFS.DOC.

After the HF step, the CIMI subroutine will localize the occupied orbitals
and assign them to domains.  For each domain a set of virtual orbitals
are created. Occupied+Virtual orbital sets are called subsystems.  A separate
input file for each subsystem is created that contains the basis set,
occupied and virtual orbitals and the AO Fock matrix.

The default procedure is that each subsystem input will be read by
the CIM subsystem driver and the correlation methods will be called.
Separate CIM drivers exist for the integral transformation and the CC
and MP2 energy codes.  These drivers include algorithms to count occupied
and virtual orbitals along with generation of the amplitude transformation
matrix that is ued to obtain the total CIM energy.  Each subsystem will also
write its own dictionary, AO and MO integral files to avoid interference
with the canonical versions of these files.  The necessary energies are
written to a small file for each subsystem and read by CIMF to compute
the total CIM energy.

The CIMI and CIMF routines come directly from Wie's original program and
have been very slightly modified.  There is no ability to run
EXETYP=CHECK in these routines and GAMESS will quit with an error
message in this case.

There are no special operations to compile CIM, it will compile
by default.  A short CIM-MP2 test can be performed after compilation with

make test EXAM_PATH=cim

Patch 2: fixing warnings and errors on other compilers
Patch 3: redorder CIMPAR and add @detail to ciminf.src

Change-Id: I1a1b653d2ff242b7a527dd186f21251b75867a6d