Bug Fix: Symmetry-adapted Coupled-cluster Normal Modes
Coupled-cluster Hessian shows a difference in the frequencies at the low frequency modes with E symmetry that are supposed to be degenerate.

Problem is caused by restarting the cluster amplitudes from the previous displacement in the numerical Hessian. Restarting the amplitudes may result different coupled-cluster energies for identical displacements (degenerate normal modes).

Change-Id: Iea8351a63e8063afcdc8ed978e41809eb552f5a0