add 10/2012 LIBCCHEM
This is the final version of Andrey Asadchev's
2012 LIBCCHEM: parallel SCF, MP2, CCSD(T)
single point energies, able to run on multiple
nodes and on GPU devices in each node.

This update is Andrey's base level LIBCCHEM.
There is an associated update to GAMESS'
scripts, installation notes, execution of
GA parallelism, and memory CHECK accompanying
this change.

Change-Id: Ic7c886c72c25621f5f4408232bd4da6438f14a66