Constrained UHF
This method places constraints on the UHF calculation
so that it converges to the ROHF energy, with no spin
contamination, and directly to the 'semi-canonical'
orbitals used in RMP perturbation after ROHF.
The procedure applies to U-HF and U-DFT energy and
gradients, and to the U-MP2 energy, where its single
single excitation term is exactly that of RMP.

Method is from
T. Tsuchimochi and G. E. Scuseria
J. Chem. Phys. 134, 164101 (2010)
Programming is from
Jason Byrd and John Montgomery

Change-Id: I40a19d00f5779bcf7f2f44efead8eb158f5096f5