LIBCHEM autotools
Avoid allowing systems with autotools 1.11 from
attempting to turn configure.ac into configure.
One must use a specially patched version of autotools
in order to have additional correctly defined macros
related to CUDA compilers to run autotools-1.11
successfully.  This change to aclocal.m4 files
ensure that regeneration does not occur, causing
the configure files from ISU to be used.

See below for a start on how to create a correctly
patched autotools-1.11:
http://code.google.com/p/asadchev/source/browse/trunk/work/autotools/?r=86
That web site is from 2010, so it may or may not reflect
all required changes to that toolkit.

Small correction to 'build-cchem' to fix a path,
and to define CUDA to point to compiler binary.

Change-Id: If7a0833fc50e908c6d9fb8776f474e7cbaedf538