Minor dipole derivative bug fix
- Added a center of mass derivative term to the dipder subroutine.
It was calculating the integral around the center of mass, but not
taking the derivative of the center of mass. That is, the integral
routine was integrating around a point, but the subroutine
used to take the derivative (derioli) wasn't including any derivatives
of the point it was integrating around. This only affects charged
molecules, since the center of mass derivative term is multiplied by
the charge. The new term is only added in when dipder is called
with CALTYP=DIPDM.

- Cleared some values in the efsp routine to avoid memory errors.

Change-Id: I295c00a794db2e4cba9d191a817f2d0a8ae44eb6