Additions to SVD project. DAF files 520-540 are used by aaron.
1.
SVD now runs with modified VVOS code from Mike's last code submission.
So, slight changes to all solutions were recorded.
2.
VVOS,SVD, and orientation codes are modified to allow for
use of ROHF in the formation of quasi-atomic orbitals.
3.
ORIKIN option was checked and was re-done to be more modular.
4.
SVD core orbitals are now formed with the high quality AOs
that are stored in the VVOS.
5.
Implement the BLOCK option for ORMAS for scftyp=mcscf.
i.e. modify NTNDEN to do what DETNO already does for
scftyp=none CI runs.

Change-Id: Ib9d0759bac4bb4eb25eacfd750125be823bc292c