Update to excorr.src and ddi_scatter_accum.c
The 1PDM is now generated from the 2PDM. This will eventually
help to eliminate much of the memory requirements for
PT2R12. exc_scatter_accum are also renamed to ddi_scatter_acc.
If excorr.src code is run in serial, the expensive sort operation
in ddi in not needed and is skipped.

Change-Id: Idfb3aae4ecdcb0e7f881313cd9f67d0a4d217a1e