diabQDPT
a) allow separate specification of the CSFs that are
considered "dominant" in the MCQDPT - this program
has somewhat different GUGA configurations than the
MCSCF.  In case the user doesn't supply such input,
the program acts exactly as before, by trying to use
the MCSCF's GUGA inputs.  Programming by Toby Zeng
as well as MWS.
b) state averaging of degenerate MCSCF states will
automatically skip the response equations, as the
symmetric SA-Lagrangian is able to deliver the correct
nuclear gradient from the non-CPMCHF code.
c) compiling script forces -m64 for older mac operating
systems, which don't default to 64 bit address space.
This was inadvertantly omitted when adapting this
script for Mavericks and gfortran 4.8.

Change-Id: I3108c7a1c000d826c524a437eda5df5dc5da9ca8